Change `setFormat` to accept `Maybe DateFormat`.
Commit 5cf41025dbbb7bba7ffc1f7861ae491a747dee6c introduced a bug which
prevented client code from being able to clear the date format in
settings. By implicitly wrapping the argument in a `Just` the actual
timestamp rendering kicked in which produced an empty field in the log
output.

This commit fixes this behaviour by changing `setFormat` to take a
`Maybe DateFormat` parameter. Client code can pass in `Nothing` to
disable timestamps.